requirements.txt
art==4.9
kafka-python==2.0.2
